fix: add new fixes to complete the DTLS pipeline

+++ Libraries/Frame/Frame.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct FrameFormat {
-    int m_timeStampMs;
+    uint64_t m_timeStampMs;
     int m_width;
     int m_height;
     int m_payloadSize;
